% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atbash.R
\name{atbash}
\alias{atbash}
\title{Encrypt or decrypt an Atbash Cipher}
\usage{
atbash(x)
}
\arguments{
\item{x}{A vector to be encoded or decoded.}
}
\value{
A character vector of length one that has been transformed
}
\description{
This can be used to create (encrypt) and solve (decrypt) an
Atbash Cipher. An Atbash Cipher swaps letters' places in the alphabet.
Thus, 'a' becomes 'z', 'b' becomes 'y', and so on. The function does not
differentiate between the two.

The Atbash Cipher Wikipedia entry provides more information on the methods
used: \url{https://en.wikipedia.org/wiki/Atbash}
}
