% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{caesars_cipher}
\alias{caesars_cipher}
\title{Function employs Caesar's cipher}
\usage{
caesars_cipher(message, nr_of_shifts, encryption)
}
\arguments{
\item{message}{string}

\item{nr_of_shifts}{int}

\item{encryption}{Either "encode" for encryption,
or "decode" for decryption.}
}
\description{
Function employs Caesar's cipher
}
\examples{
# Encode your message
caesars_cipher("theeaglehaslanden", 4, encryption = "encode")

# Decode your received message
caesars_cipher("XLIIEKPILEWPERHIR", 4, encryption = "decode")
}
