% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningKeyCipher_final_commented_PrepText.R
\name{decipher}
\alias{decipher}
\title{A Running Key Encipher Function}
\usage{
decipher()
}
\description{
This function deciphers text using a Running Key Cipher.
}
\examples{
decipher()
}
\keyword{cipher}
\keyword{key}
\keyword{running}
