% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeaserCipher.R
\name{decipher_c}
\alias{decipher_c}
\title{A Caeser Cipher Function}
\usage{
decipher_c()
}
\description{
This function deciphers text using a Caeser Cipher.
}
\examples{
decipher_c()
}
\keyword{Caeser}
\keyword{cipher}
