% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningKeyCipher_final_commented_PrepText.R
\name{encipher}
\alias{encipher}
\title{A Running Key Encipher Function}
\usage{
encipher()
}
\description{
This function enciphers text using a Running Key Cipher.
}
\examples{
encipher()
}
\keyword{cipher}
\keyword{key}
\keyword{running}
