% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeaserCipher.R
\name{encipher_c}
\alias{encipher_c}
\title{A Caeser Cipher Function}
\usage{
encipher_c()
}
\description{
This function enciphers text using a Caeser Cipher.
}
\examples{
encipher_c()
}
\keyword{Caeser}
\keyword{cipher}
