% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeaserCipher.R
\name{prep.text}
\alias{prep.text}
\title{A Text Preparation Function}
\usage{
prep.text()
}
\description{
This function prepares text for encryption by removing spaces and special characters.
}
\examples{
prep.text()
}
\keyword{preparation}
\keyword{text}
