% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenere.R
\name{random_key_gen}
\alias{random_key_gen}
\title{Produce a key for encryption purposes}
\usage{
random_key_gen(length_keyword)
}
\arguments{
\item{length_keyword}{int}
}
\value{
numeric
}
\description{
Produce a key for encryption purposes
}
\examples{
# Generate a random key to make Le Chiffre Indechiffrable
key <- random_key_gen(nchar("THEEAGLEHASLANDED"))
}
