% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenere.R
\name{vigenere_cipher}
\alias{vigenere_cipher}
\title{Use the Vigenere cipher.}
\usage{
vigenere_cipher(plaintext, keyword, encryption = "encode")
}
\arguments{
\item{plaintext}{A string. Your message to encode/decode.}

\item{keyword}{string. The secret keyword.}

\item{encryption}{Either "encode" "or decode".}
}
\description{
Use the Vigenere cipher.
}
\examples{
#' # Encode your message
vigenere_cipher(plaintext = "diverttroopstoeastridge", keyword = "WHITE", encryption = "encode")
# Decode your message
vigenere_cipher(plaintext = "ZPDXVPAZHSLZBHIWZBKMZNM", keyword = "WHITE", encryption = "decode")
}
